Set up routes, create profile page
